requirements.txt
procyclingstats
requests
python-dotenv
rich
beautifulsoup4
